LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/custom-elements/builtin-coverage-expected.txt


FAIL a: Define a customized built-in element The superclass is not a constructor.
PASS abbr: Define a customized built-in element 
PASS abbr: Operator 'new' should instantiate a customized built-in element 
FAIL abbr: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL abbr: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
PASS address: Define a customized built-in element 
PASS address: Operator 'new' should instantiate a customized built-in element 
FAIL address: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL address: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL area: Define a customized built-in element The superclass is not a constructor.
PASS article: Define a customized built-in element 
PASS article: Operator 'new' should instantiate a customized built-in element 
FAIL article: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL article: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
PASS aside: Define a customized built-in element 
PASS aside: Operator 'new' should instantiate a customized built-in element 
FAIL aside: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL aside: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL audio: Define a customized built-in element The superclass is not a constructor.
PASS b: Define a customized built-in element 
PASS b: Operator 'new' should instantiate a customized built-in element 
FAIL b: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL b: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL base: Define a customized built-in element The superclass is not a constructor.
PASS bdi: Define a customized built-in element 
PASS bdi: Operator 'new' should instantiate a customized built-in element 
FAIL bdi: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL bdi: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
PASS bdo: Define a customized built-in element 
PASS bdo: Operator 'new' should instantiate a customized built-in element 
FAIL bdo: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL bdo: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL blockquote: Define a customized built-in element The superclass is not a constructor.
FAIL body: Define a customized built-in element The superclass is not a constructor.
FAIL br: Define a customized built-in element The superclass is not a constructor.
FAIL button: Define a customized built-in element The superclass is not a constructor.
FAIL canvas: Define a customized built-in element The superclass is not a constructor.
FAIL caption: Define a customized built-in element The superclass is not a constructor.
PASS cite: Define a customized built-in element 
PASS cite: Operator 'new' should instantiate a customized built-in element 
FAIL cite: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL cite: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
PASS code: Define a customized built-in element 
PASS code: Operator 'new' should instantiate a customized built-in element 
FAIL code: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL code: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL col: Define a customized built-in element The superclass is not a constructor.
FAIL colgroup: Define a customized built-in element The superclass is not a constructor.
FAIL data: Define a customized built-in element The superclass is not a constructor.
PASS dd: Define a customized built-in element 
PASS dd: Operator 'new' should instantiate a customized built-in element 
FAIL dd: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL dd: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL del: Define a customized built-in element The superclass is not a constructor.
FAIL details: Define a customized built-in element The superclass is not a constructor.
PASS dfn: Define a customized built-in element 
PASS dfn: Operator 'new' should instantiate a customized built-in element 
FAIL dfn: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL dfn: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL div: Define a customized built-in element The superclass is not a constructor.
FAIL dl: Define a customized built-in element The superclass is not a constructor.
PASS dt: Define a customized built-in element 
PASS dt: Operator 'new' should instantiate a customized built-in element 
FAIL dt: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL dt: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
PASS em: Define a customized built-in element 
PASS em: Operator 'new' should instantiate a customized built-in element 
FAIL em: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL em: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL embed: Define a customized built-in element The superclass is not a constructor.
FAIL fieldset: Define a customized built-in element The superclass is not a constructor.
PASS figcaption: Define a customized built-in element 
PASS figcaption: Operator 'new' should instantiate a customized built-in element 
FAIL figcaption: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL figcaption: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
PASS figure: Define a customized built-in element 
PASS figure: Operator 'new' should instantiate a customized built-in element 
FAIL figure: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL figure: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
PASS footer: Define a customized built-in element 
PASS footer: Operator 'new' should instantiate a customized built-in element 
FAIL footer: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL footer: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL form: Define a customized built-in element The superclass is not a constructor.
FAIL h1: Define a customized built-in element The superclass is not a constructor.
FAIL h2: Define a customized built-in element The superclass is not a constructor.
FAIL h3: Define a customized built-in element The superclass is not a constructor.
FAIL h4: Define a customized built-in element The superclass is not a constructor.
FAIL h5: Define a customized built-in element The superclass is not a constructor.
FAIL h6: Define a customized built-in element The superclass is not a constructor.
PASS header: Define a customized built-in element 
PASS header: Operator 'new' should instantiate a customized built-in element 
FAIL header: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL header: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
PASS hgroup: Define a customized built-in element 
PASS hgroup: Operator 'new' should instantiate a customized built-in element 
FAIL hgroup: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL hgroup: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL hr: Define a customized built-in element The superclass is not a constructor.
FAIL html: Define a customized built-in element The superclass is not a constructor.
PASS i: Define a customized built-in element 
PASS i: Operator 'new' should instantiate a customized built-in element 
FAIL i: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL i: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL iframe: Define a customized built-in element The superclass is not a constructor.
FAIL img: Define a customized built-in element The superclass is not a constructor.
FAIL input: Define a customized built-in element The superclass is not a constructor.
FAIL ins: Define a customized built-in element The superclass is not a constructor.
PASS kbd: Define a customized built-in element 
PASS kbd: Operator 'new' should instantiate a customized built-in element 
FAIL kbd: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL kbd: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL label: Define a customized built-in element The superclass is not a constructor.
FAIL legend: Define a customized built-in element The superclass is not a constructor.
FAIL li: Define a customized built-in element The superclass is not a constructor.
FAIL link: Define a customized built-in element The superclass is not a constructor.
PASS main: Define a customized built-in element 
PASS main: Operator 'new' should instantiate a customized built-in element 
FAIL main: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL main: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL map: Define a customized built-in element The superclass is not a constructor.
PASS mark: Define a customized built-in element 
PASS mark: Operator 'new' should instantiate a customized built-in element 
FAIL mark: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL mark: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL menu: Define a customized built-in element The superclass is not a constructor.
FAIL meta: Define a customized built-in element The superclass is not a constructor.
FAIL meter: Define a customized built-in element The superclass is not a constructor.
PASS nav: Define a customized built-in element 
PASS nav: Operator 'new' should instantiate a customized built-in element 
FAIL nav: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL nav: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
PASS noscript: Define a customized built-in element 
PASS noscript: Operator 'new' should instantiate a customized built-in element 
FAIL noscript: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL noscript: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL object: Define a customized built-in element The superclass is not a constructor.
FAIL ol: Define a customized built-in element The superclass is not a constructor.
FAIL optgroup: Define a customized built-in element The superclass is not a constructor.
FAIL option: Define a customized built-in element The superclass is not a constructor.
FAIL output: Define a customized built-in element The superclass is not a constructor.
FAIL p: Define a customized built-in element The superclass is not a constructor.
FAIL param: Define a customized built-in element The superclass is not a constructor.
FAIL picture: Define a customized built-in element The superclass is not a constructor.
FAIL pre: Define a customized built-in element The superclass is not a constructor.
FAIL progress: Define a customized built-in element The superclass is not a constructor.
FAIL q: Define a customized built-in element The superclass is not a constructor.
PASS rp: Define a customized built-in element 
PASS rp: Operator 'new' should instantiate a customized built-in element 
FAIL rp: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL rp: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
PASS rt: Define a customized built-in element 
PASS rt: Operator 'new' should instantiate a customized built-in element 
FAIL rt: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL rt: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
PASS ruby: Define a customized built-in element 
PASS ruby: Operator 'new' should instantiate a customized built-in element 
FAIL ruby: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL ruby: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
PASS s: Define a customized built-in element 
PASS s: Operator 'new' should instantiate a customized built-in element 
FAIL s: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL s: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
PASS samp: Define a customized built-in element 
PASS samp: Operator 'new' should instantiate a customized built-in element 
FAIL samp: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL samp: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL script: Define a customized built-in element The superclass is not a constructor.
PASS section: Define a customized built-in element 
PASS section: Operator 'new' should instantiate a customized built-in element 
FAIL section: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL section: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL select: Define a customized built-in element The superclass is not a constructor.
PASS small: Define a customized built-in element 
PASS small: Operator 'new' should instantiate a customized built-in element 
FAIL small: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL small: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL source: Define a customized built-in element The superclass is not a constructor.
FAIL span: Define a customized built-in element The superclass is not a constructor.
PASS strong: Define a customized built-in element 
PASS strong: Operator 'new' should instantiate a customized built-in element 
FAIL strong: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL strong: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL style: Define a customized built-in element The superclass is not a constructor.
PASS sub: Define a customized built-in element 
PASS sub: Operator 'new' should instantiate a customized built-in element 
FAIL sub: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL sub: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
PASS summary: Define a customized built-in element 
PASS summary: Operator 'new' should instantiate a customized built-in element 
FAIL summary: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL summary: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
PASS sup: Define a customized built-in element 
PASS sup: Operator 'new' should instantiate a customized built-in element 
FAIL sup: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL sup: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL table: Define a customized built-in element The superclass is not a constructor.
FAIL tbody: Define a customized built-in element The superclass is not a constructor.
FAIL td: Define a customized built-in element The superclass is not a constructor.
FAIL template: Define a customized built-in element The superclass is not a constructor.
FAIL textarea: Define a customized built-in element The superclass is not a constructor.
FAIL tfoot: Define a customized built-in element The superclass is not a constructor.
FAIL th: Define a customized built-in element The superclass is not a constructor.
FAIL thead: Define a customized built-in element The superclass is not a constructor.
FAIL time: Define a customized built-in element The superclass is not a constructor.
FAIL title: Define a customized built-in element The superclass is not a constructor.
FAIL tr: Define a customized built-in element The superclass is not a constructor.
FAIL track: Define a customized built-in element The superclass is not a constructor.
PASS u: Define a customized built-in element 
PASS u: Operator 'new' should instantiate a customized built-in element 
FAIL u: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL u: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL ul: Define a customized built-in element The superclass is not a constructor.
PASS var: Define a customized built-in element 
PASS var: Operator 'new' should instantiate a customized built-in element 
FAIL var: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL var: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL video: Define a customized built-in element The superclass is not a constructor.
PASS wbr: Define a customized built-in element 
PASS wbr: Operator 'new' should instantiate a customized built-in element 
FAIL wbr: document.createElement() should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL wbr: innerHTML should instantiate a customized built-in element assert_equals: expected function "class extends HTMLElement {}" but got function "function HTMLElement() {
    [native code]
}"
FAIL datalist: Define a customized built-in element The superclass is not a constructor.
FAIL dialog: Define a customized built-in element The superclass is not a constructor.
FAIL slot: Define a customized built-in element The superclass is not a constructor.
